// src/generated/Preview.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Preview = { "type": "text", text: string, } | { "type": "image", id: number, width: number, height: number, } | { "type": "files", paths: Array<string>, } | { "type": "file", path: string, size: number, modified: number | null, isDir: boolean, };